2.0.2 and slightly beyond.
Docs: add warning about crashing if ‘executable’ is not.
Executable that is.

Make AsyncShellTask a class.
It only contains objects so everything should continue working as before. It is basically a wrapper for NSTask so it should also be a class.

Readme: add section about error handling from commands.

Improve documentation for what "runAndPrint" throws.