Massive renaming: avoid words Type, Task and Shell.
Protocols should no longer end in Type, Task has been renamed in Foundation and the word Shell is not appropriate for running commands / launching processes, it describes terminal applications and shell interpreters like bash. Which means the name SwiftShell itself is probably a misnomer but I am not about to change that now.

Also remove any mentions of SwiftShell 2.